+++ src/Exceptions/InvalidSentenceException.php
@@ -1,0 +1,16 @@
+<?php
+/*
+ * This file is part of the Brackets Checker library.
+ *
+ * (c) Max Voronov <maxivoronov@gmail.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace MaxVoronov\BracketsChecker\Exceptions;
+
+class InvalidSentenceException extends \InvalidArgumentException
+{
+    //
+}