+++ CHANGELOG.md
@@ -1,3 +1,12 @@
 # Changelog
 
 All Notable changes to `brackets-checker` library will be documented in this file.
+
+## 1.0.0
+- Improved library API
+- Added checking custom brackets
+- Added custom filtering characters
+- Extended library exceptions
+
+## 0.1.0
+- Initial version